+++ mlir/lib/Conversion/MathToEmitC/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_mlir_conversion_library(MLIRMathToEmitC
   MathToEmitC.cpp
+  MathToEmitCPass.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${MLIR_MAIN_INCLUDE_DIR}/mlir/Conversion/MathToEmitC
